[Refactor] Split elementwise kernel support modules
