+++ src/tileops/kernels/elementwise/_base.py
@@ -25,16 +25,41 @@ follows from the format: e4m3fn has no Inf, so a saturating ``T.Cast`` clamping 
 ±448.0 is what it can represent; e5m2 does have Inf, so the PrimFunc emits fp16 and
 ``forward`` casts to e5m2 without saturating, which keeps Inf and NaN."""
 
-import functools
 import math
-import warnings
-from dataclasses import dataclass
 
-import tilelang
 import tilelang.language as T
 import torch
 
 from tileops.kernels.kernel_base import Kernel
+
+from ._broadcast import _broadcast_target as _broadcast_target
+from ._broadcast import _expand_flat as _expand_flat
+from ._broadcast import _flat, _is_contiguous_same_shape, coalesce_broadcast_dims
+from ._builders import (
+    _make_binary_direct,
+    _make_binary_explicit,
+    _make_binary_register_copy,
+    _make_fused_gated_direct,
+    _make_fused_gated_explicit,
+    _make_unary_direct,
+    _make_unary_explicit,
+    _make_unary_regcopy,
+)
+from ._dtype import _BINARY_FULL_DTYPES as _BINARY_FULL_DTYPES
+from ._dtype import _BINARY_NO_BOOL_DTYPES as _BINARY_NO_BOOL_DTYPES
+from ._dtype import _BITWISE_DTYPES, _FLOAT_DTYPES, _LOGICAL_DTYPES, _is_fp8
+from ._dtype import _FP8_DTYPES as _FP8_DTYPES
+from ._dtype import _clamp_to_dtype_range as _clamp_to_dtype_range
+from ._dtype import _fp8_accum_dtype_str as _fp8_accum_dtype_str
+from ._launch import _DEFAULT_THREADS, _LAUNCH_POLICY
+from ._output import (
+    ElementwiseOutputPlan,
+    _get_fp8_output_dtypes,
+    _store_binary_bool_as_int8,
+    _store_unary_bool_as_int8,
+    _wrap_fp8_accumulation,
+)
+from ._strategy import BinaryStrategyPolicy, UnaryStrategyPolicy
 
 __all__ = [
     "BinaryKernel",
@@ -47,387 +72,6 @@ __all__ = [
     "coalesce_broadcast_dims",
 ]
 
-_BITWISE_DTYPES = (
-    torch.bool,
-    torch.uint8,
-    torch.int8,
-    torch.int16,
-    torch.int32,
-    torch.int64,
-)
-
-
-_FP8_DTYPES = (
-    torch.float8_e4m3fn,
-    torch.float8_e5m2,
-)
-
-
-_FLOAT_DTYPES = (
-    torch.float16,
-    torch.bfloat16,
-    torch.float32,
-)
-
-
-_LOGICAL_DTYPES = _BITWISE_DTYPES + _FLOAT_DTYPES
-
-
-_BINARY_FULL_DTYPES = _BITWISE_DTYPES + (
-    torch.float16,
-    torch.bfloat16,
-    torch.float32,
-)
-
-
-_BINARY_NO_BOOL_DTYPES = tuple(dt for dt in _BINARY_FULL_DTYPES if dt is not torch.bool)
-
-
-def _is_fp8(dtype: torch.dtype) -> bool:
-    """Check if a torch dtype is an fp8 variant."""
-    return dtype in _FP8_DTYPES
-
-
-def _torch_dtype_nbytes(dtype: torch.dtype) -> int:
-    """Return the byte width of a torch dtype."""
-    return torch.empty(0, dtype=dtype).element_size()
-
-
-#: What a bool result is written through where the kernel can choose. int8 is the same
-#: width as bool and 0/1 are the same two byte patterns, so the bool view copies nothing.
-_BOOL_STORAGE_DTYPE: str = "int8"
-
-
-def _store_unary_bool_as_int8(op_func):
-    """Wrap a unary predicate so its bool result lands as ``_BOOL_STORAGE_DTYPE``."""
-
-    def wrapped(x):
-        return T.if_then_else(
-            op_func(x),
-            T.cast(1, _BOOL_STORAGE_DTYPE),
-            T.cast(0, _BOOL_STORAGE_DTYPE),
-        )
-
-    return wrapped
-
-
-def _store_binary_bool_as_int8(op_func):
-    """Wrap a binary predicate so its bool result lands as ``_BOOL_STORAGE_DTYPE``."""
-
-    def wrapped(a, b):
-        return T.if_then_else(
-            op_func(a, b),
-            T.cast(1, _BOOL_STORAGE_DTYPE),
-            T.cast(0, _BOOL_STORAGE_DTYPE),
-        )
-
-    return wrapped
-
-
-@dataclass(frozen=True)
-class ElementwiseLaunchConfig:
-    """Concrete launch knobs for one elementwise specialization."""
-
-    strategy: str
-    threads: int
-    num_per_thread: int
-
-    def as_dict(self) -> dict:
-        """Return the config shape expected by ``Kernel.init_config``."""
-        return {
-            "strategy": self.strategy,
-            "threads": self.threads,
-            "num_per_thread": self.num_per_thread,
-        }
-
-
-@dataclass(frozen=True)
-class ElementwiseLaunchPolicy:
-    """Heuristics that turn element dtype and shape into launch knobs."""
-
-    default_threads: int = 128
-    bytes_per_thread: int = 16
-    min_num_per_thread: int = 4
-    bool_output_max_num_per_thread: int = 4
-    max_threads: int = 1024
-    target_blocks: int = 256
-    fp8_num_per_thread: int = 16
-
-    def default_config(
-        self,
-        *,
-        strategy: str,
-        input_dtype: torch.dtype,
-        output_dtype: torch.dtype,
-        n_total: int | None,
-        stores_bool: bool = True,
-    ) -> ElementwiseLaunchConfig:
-        """Return the default launch config for this specialization.
-
-        The rule order is significant: fp8 keeps its historical 128-bit default;
-        otherwise start from input width, cap real bool stores, widen for narrow
-        non-bool outputs, then shrink small non-direct launches.
-        """
-        if _is_fp8(input_dtype):
-            return ElementwiseLaunchConfig(
-                strategy=strategy,
-                threads=self.default_threads,
-                num_per_thread=self.fp8_num_per_thread,
-            )
-        threads = self.default_threads
-        npt = self._base_num_per_thread(input_dtype)
-        threads, npt, handled_bool = self._adjust_for_bool_output(
-            threads,
-            npt,
-            output_dtype,
-            stores_bool,
-        )
-        if not handled_bool:
-            npt = self._adjust_for_narrow_output(npt, input_dtype, output_dtype)
-        npt = self._shrink_npt_for_small_tensor(strategy, n_total, threads, npt)
-        return ElementwiseLaunchConfig(strategy=strategy, threads=threads, num_per_thread=npt)
-
-    def _base_num_per_thread(self, dtype: torch.dtype) -> int:
-        elem_bytes = _torch_dtype_nbytes(dtype)
-        return max(self.min_num_per_thread, self.bytes_per_thread // elem_bytes)
-
-    def _adjust_for_bool_output(
-        self,
-        threads: int,
-        npt: int,
-        output_dtype: torch.dtype,
-        stores_bool: bool,
-    ) -> tuple[int, int, bool]:
-        if output_dtype != torch.bool or not stores_bool:
-            return threads, npt, False
-        capped = min(npt, self.bool_output_max_num_per_thread)
-        return min(self.max_threads, threads * npt // capped), capped, True
-
-    def _adjust_for_narrow_output(
-        self,
-        npt: int,
-        input_dtype: torch.dtype,
-        output_dtype: torch.dtype,
-    ) -> int:
-        elem_bytes = _torch_dtype_nbytes(input_dtype)
-        stored_bytes = _torch_dtype_nbytes(output_dtype)
-        return npt * 2 if stored_bytes < elem_bytes else npt
-
-    def _shrink_npt_for_small_tensor(
-        self,
-        strategy: str,
-        n_total: int | None,
-        threads: int,
-        npt: int,
-    ) -> int:
-        if n_total is None or strategy == "direct":
-            return npt
-        while npt > self.min_num_per_thread and n_total < threads * npt * self.target_blocks:
-            npt //= 2
-        return npt
-
-
-_LAUNCH_POLICY = ElementwiseLaunchPolicy()
-_DEFAULT_THREADS: int = _LAUNCH_POLICY.default_threads
-
-
-@dataclass(frozen=True)
-class ElementwiseOutputPlan:
-    """Storage decisions for one elementwise kernel specialization."""
-
-    logical_dtype: torch.dtype
-    kernel_output_dtype: str | None
-    post_cast_dtype: torch.dtype | None = None
-    bool_via_int8: bool = False
-
-    @classmethod
-    def for_unary(
-        cls,
-        input_dtype: torch.dtype,
-        declared_output_dtype: torch.dtype | None,
-        strategy: str,
-    ) -> "ElementwiseOutputPlan":
-        logical_dtype, post_cast_dtype = _logical_output_dtype(input_dtype, declared_output_dtype)
-        bool_via_int8 = declared_output_dtype == torch.bool and strategy == "register_copy"
-        return cls._from_parts(logical_dtype, post_cast_dtype, bool_via_int8)
-
-    @classmethod
-    def for_binary(
-        cls,
-        input_dtype: torch.dtype,
-        declared_output_dtype: torch.dtype | None,
-        strategy: str,
-    ) -> "ElementwiseOutputPlan":
-        logical_dtype, post_cast_dtype = _logical_output_dtype(input_dtype, declared_output_dtype)
-        bool_via_int8 = declared_output_dtype == torch.bool and strategy == "register_copy"
-        return cls._from_parts(logical_dtype, post_cast_dtype, bool_via_int8)
-
-    @classmethod
-    def for_fused_gated(cls, input_dtype: torch.dtype) -> "ElementwiseOutputPlan":
-        logical_dtype, post_cast_dtype = _logical_output_dtype(input_dtype, None)
-        return cls._from_parts(logical_dtype, post_cast_dtype, bool_via_int8=False)
-
-    @classmethod
-    def _from_parts(
-        cls,
-        logical_dtype: torch.dtype,
-        post_cast_dtype: torch.dtype | None,
-        bool_via_int8: bool,
-    ) -> "ElementwiseOutputPlan":
-        if bool_via_int8:
-            kernel_output_dtype = _BOOL_STORAGE_DTYPE
-        elif post_cast_dtype is not None:
-            kernel_output_dtype = _fp8_accum_dtype_str()
-        else:
-            kernel_output_dtype = Kernel.dtype_to_str(logical_dtype)
-        return cls(
-            logical_dtype=logical_dtype,
-            kernel_output_dtype=kernel_output_dtype,
-            post_cast_dtype=post_cast_dtype,
-            bool_via_int8=bool_via_int8,
-        )
-
-
-def _logical_output_dtype(
-    input_dtype: torch.dtype,
-    declared_output_dtype: torch.dtype | None,
-) -> tuple[torch.dtype, torch.dtype | None]:
-    """Return the logical output dtype and optional post-cast dtype."""
-    if (
-        declared_output_dtype is None
-        and _is_fp8(input_dtype)
-        and _fp8_needs_nonsaturating_cast(input_dtype)
-    ):
-        return torch.float16, input_dtype
-    return declared_output_dtype or input_dtype, None
-
-
-def _bool_output_needs_scalar(
-    input_dtype: torch.dtype,
-    declared_output_dtype: torch.dtype | None,
-) -> bool:
-    """Whether TileLang cannot lower this bool-output vectorized path."""
-    return declared_output_dtype == torch.bool and input_dtype in (
-        torch.uint8,
-        torch.int8,
-        torch.int16,
-    )
-
-
-@dataclass(frozen=True)
-class UnaryStrategyPolicy:
-    """Strategy selection rules for one-input elementwise kernels."""
-
-    strategies: list[str]
-    default_strategy: str
-
-    def choose(
-        self,
-        *,
-        requested: str | None,
-        input_dtype: torch.dtype,
-        declared_output_dtype: torch.dtype | None,
-    ) -> str:
-        self._validate_requested(requested)
-        if input_dtype == torch.bool:
-            self._warn_direct_override(
-                requested,
-                "UnaryKernel: dtype=torch.bool requires strategy="
-                "'direct' (TileLang cannot lower vectorised boolx<N> "
-                "loads); overriding requested strategy={requested!r}.",
-            )
-            return "direct"
-        if _bool_output_needs_scalar(input_dtype, declared_output_dtype):
-            self._warn_direct_override(
-                requested,
-                "UnaryKernel: dtype={dtype} with torch.bool output "
-                "requires strategy='direct' (TileLang cannot lower "
-                "vectorised boolx<N> stores for sub-32-bit integer "
-                "inputs); overriding requested strategy={requested!r}.",
-                dtype=input_dtype,
-            )
-            return "direct"
-        if requested is None and _is_fp8(input_dtype):
-            return "explicit_parallel"
-        return requested or self.default_strategy
-
-    def _validate_requested(self, requested: str | None) -> None:
-        if requested is not None and requested not in self.strategies:
-            raise ValueError(f"Unknown strategy '{requested}', expected one of {self.strategies}")
-
-    @staticmethod
-    def _warn_direct_override(
-        requested: str | None,
-        template: str,
-        **kwargs,
-    ) -> None:
-        if requested is None or requested == "direct":
-            return
-        warnings.warn(
-            template.format(requested=requested, **kwargs),
-            RuntimeWarning,
-            stacklevel=3,
-        )
-
-
-@dataclass(frozen=True)
-class BinaryStrategyPolicy:
-    """Strategy selection rules for broadcast-capable binary elementwise kernels."""
-
-    strategies: list[str]
-    default_strategy: str
-
-    def choose(
-        self,
-        *,
-        requested: str | None,
-        input_dtype: torch.dtype,
-        declared_output_dtype: torch.dtype | None,
-        same_shape: bool,
-    ) -> str:
-        self._validate_requested(requested)
-        if input_dtype == torch.bool:
-            self._warn_direct_override(
-                requested,
-                "BinaryKernel: dtype=torch.bool requires strategy="
-                "'direct' (TileLang cannot lower vectorised boolx<N> "
-                "loads); overriding requested strategy={requested!r}.",
-            )
-            return "direct"
-        if _bool_output_needs_scalar(input_dtype, declared_output_dtype):
-            self._warn_direct_override(
-                requested,
-                "BinaryKernel: dtype={dtype} with torch.bool output "
-                "requires strategy='direct' (TileLang cannot lower "
-                "vectorised boolx<N> stores for sub-32-bit integer "
-                "inputs); overriding requested strategy={requested!r}.",
-                dtype=input_dtype,
-            )
-            return "direct"
-        if requested is not None:
-            if requested == "register_copy" and not same_shape:
-                return "explicit_parallel"
-            return requested
-        return "register_copy" if same_shape else self.default_strategy
-
-    def _validate_requested(self, requested: str | None) -> None:
-        if requested is not None and requested not in self.strategies:
-            raise ValueError(f"Unknown strategy '{requested}', expected one of {self.strategies}")
-
-    @staticmethod
-    def _warn_direct_override(
-        requested: str | None,
-        template: str,
-        **kwargs,
-    ) -> None:
-        if requested is None or requested == "direct":
-            return
-        warnings.warn(
-            template.format(requested=requested, **kwargs),
-            RuntimeWarning,
-            stacklevel=3,
-        )
-
 
 def _log_for_output_precision(value, wide):
     """Return ``log(wide)`` computed to the precision *value*'s dtype can keep.
@@ -435,562 +79,6 @@ def _log_for_output_precision(value, wide):
     A narrow result cannot keep the precision difference, so it takes the fast log.
     """
     return T.log(wide) if value.dtype == "float32" else T.__log(wide)
-
-
-def _fp8_needs_nonsaturating_cast(dtype: torch.dtype) -> bool:
-    """Return True if the fp8 format supports Inf/NaN and needs non-saturating output.
-
-    e5m2 has Inf/NaN representation -- TileLang's T.Cast uses saturating conversion
-    which incorrectly clamps Inf to max-finite.  For e5m2, the kernel must produce
-    fp16 output and let PyTorch do the final non-saturating cast.
-
-    e4m3fn has no Inf representation, so saturating T.Cast is correct.
-    """
-    return dtype == torch.float8_e5m2
-
-
-def _fp8_accum_dtype_str() -> str:
-    """Return the TileLang dtype string used for fp8 intermediate accumulation."""
-    return "float16"
-
-
-def _get_fp8_output_dtypes(dtype: torch.dtype):
-    """Return (fp8_output_dtype, kernel_output_dtype) for fp8 handling.
-
-    For e5m2: the PrimFunc produces fp16 to preserve Inf/NaN, and ``forward`` casts
-    to e5m2 without saturating.
-    For e4m3fn or non-fp8: the PrimFunc outputs the input dtype directly.
-
-    Returns:
-        Tuple of (_fp8_output_dtype, output_dtype).  _fp8_output_dtype is
-        the original fp8 dtype when a post-cast is needed, else None.
-    """
-    if _is_fp8(dtype) and _fp8_needs_nonsaturating_cast(dtype):
-        return dtype, torch.float16
-    return None, dtype
-
-
-def _clamp_to_dtype_range(value, dtype: torch.dtype):
-    """Normalize *value* into the storage representation of *dtype*.
-
-    Mirrors PyTorch ``Tensor.masked_fill`` scalar coercion so the literal lands
-    as the same bit pattern PyTorch would write:
-
-    - bool: non-zero → ``1``, else ``0``.
-    - Signed int: truncate toward zero; ``+/-Inf`` maps to ``iinfo.max/min``
-      so a bypassed validator cannot raise ``OverflowError`` on ``int(inf)``.
-    - ``uint8``: negatives wrap via ``& 0xFF``, non-negatives truncate.
-    - ``fp16/bf16/fp32`` and ``fp8_e5m2``: ``NaN`` / ``+-Inf`` pass through,
-      finite values clamp to ``finfo``.
-    - ``fp8_e4m3fn`` has no Inf, so ``+-Inf`` saturates to ``finfo.max/min``
-      to avoid a TVM ``FloatImm`` overflow.
-    """
-    if dtype == torch.bool:
-        return 1 if bool(value) else 0
-    if dtype in _BITWISE_DTYPES:
-        if isinstance(value, float) and math.isinf(value):
-            iinfo = torch.iinfo(dtype)
-            return iinfo.max if value > 0 else iinfo.min
-        if (
-            dtype == torch.uint8
-            and isinstance(value, int)
-            and not isinstance(value, bool)
-            and value < 0
-        ):
-            return value & 0xFF
-        return int(value)
-    fvalue = float(value)
-    if math.isnan(fvalue):
-        return fvalue
-    finfo = torch.finfo(dtype)
-    if math.isinf(fvalue):
-        if dtype in _FP8_DTYPES and not _fp8_needs_nonsaturating_cast(dtype):
-            return finfo.max if fvalue > 0 else finfo.min
-        return fvalue
-    return max(finfo.min, min(finfo.max, fvalue))
-
-
-def _wrap_fp8_accumulation(base_op, dtype, dtype_str, arity=1):
-    """Wrap an op function with fp8 accumulation logic if *dtype* is fp8.
-
-    Both fp8 dtypes cast inputs to fp16 and compute there. e4m3fn casts the
-    result back via saturating ``T.Cast`` (correct — it has no Inf); e5m2
-    leaves the result in fp16 and ``forward`` does the final non-saturating
-    cast, which preserves Inf/NaN.
-
-    Non-fp8 dtypes get *base_op* back unchanged.
-    """
-    if not _is_fp8(dtype):
-        return base_op
-
-    accum = _fp8_accum_dtype_str()
-
-    if _fp8_needs_nonsaturating_cast(dtype):
-        # e5m2: compute in fp16, leave result as fp16
-        if arity == 1:
-
-            def fp8_accum_op(x):
-                return base_op(T.cast(x, accum))
-        else:
-
-            def fp8_accum_op(a, b):
-                return base_op(T.cast(a, accum), T.cast(b, accum))
-
-        return fp8_accum_op
-
-    # e4m3fn: compute in fp16, saturating cast back
-    if arity == 1:
-
-        def fp8_accum_op(x):
-            return T.Cast(dtype_str, base_op(T.cast(x, accum)))
-    else:
-
-        def fp8_accum_op(a, b):
-            return T.Cast(dtype_str, base_op(T.cast(a, accum), T.cast(b, accum)))
-
-    return fp8_accum_op
-
-
-@functools.lru_cache(maxsize=32)
-def _make_unary_direct(N, dtype, op_func, output_dtype=None, threads=256):
-    """Strategy 1: 1 element per thread."""
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[1])
-    def kernel(threads_arg):
-        @T.prim_func
-        def main(x: T.Tensor((N,), dtype), y: T.Tensor((N,), out_dtype)):
-            with T.Kernel(T.ceildiv(N, threads_arg), threads=threads_arg) as bx:
-                for i in T.Parallel(threads_arg):
-                    idx = bx * threads_arg + i
-                    y[idx] = op_func(x[idx])
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_unary_explicit(N, dtype, op_func, output_dtype=None, threads=256, num_per_thread=8):
-    """Strategy 2: N elements per thread via T.Parallel(threads, npt).
-
-    The grid comes from the specialized arguments, not from *threads* and
-    *num_per_thread*: those name the maker's cache entry, and a call indexing a
-    narrower block than the grid was sized for leaves part of the tensor unwritten.
-    """
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[1])
-    def kernel(threads_arg, npt_arg):
-        block_size = threads_arg * npt_arg
-
-        @T.prim_func
-        def main(x: T.Tensor((N,), dtype), y: T.Tensor((N,), out_dtype)):
-            with T.Kernel(T.ceildiv(N, block_size), threads=threads_arg) as bx:
-                for i, j in T.Parallel(threads_arg, npt_arg):
-                    idx = (bx * threads_arg + i) * npt_arg + j
-                    y[idx] = op_func(x[idx])
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_unary_regcopy(N, dtype, op_func, output_dtype=None, threads=256, num_per_thread=8):
-    """Strategy 3: fragment load → compute → fragment store.
-
-    Grid and fragment extent come from the specialized arguments, for the reason
-    :func:`_make_unary_explicit` gives.
-    """
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[1])
-    def kernel(threads_arg, npt_arg):
-        block_size = threads_arg * npt_arg
-
-        @T.prim_func
-        def main(x: T.Tensor((N,), dtype), y: T.Tensor((N,), out_dtype)):
-            with T.Kernel(T.ceildiv(N, block_size), threads=threads_arg) as bx:
-                x_reg = T.alloc_fragment((block_size,), dtype)
-                y_reg = T.alloc_fragment((block_size,), out_dtype)
-                T.copy(x[bx * block_size : (bx + 1) * block_size], x_reg)
-                for i, j in T.Parallel(threads_arg, npt_arg):
-                    y_reg[i * npt_arg + j] = op_func(x_reg[i * npt_arg + j])
-                T.copy(y_reg, y[bx * block_size : (bx + 1) * block_size])
-
-        return main
-
-    return kernel
-
-
-def _flat(t):
-    """The flat view every PrimFunc here takes.
-
-    The Op layer normalizes contiguity and hands over the manifest-declared
-    shape, so reducing it to one dimension is this backend's business.
-    """
-    return t.reshape(-1)
-
-
-def _broadcast_target(*tensors):
-    """The output shape a multi-operand kernel writes, from the operands it got.
-
-    What ``torch.broadcast_shapes`` says, which is what the manifest's shape rules say:
-    all-0-dim operands give ``()``. The PrimFuncs below work on a flat buffer either
-    way — a 0-dim tensor holds one element — so restoring that shape is this wrapper's
-    last step.
-    """
-    return torch.broadcast_shapes(*(tuple(t.shape) for t in tensors if t is not None))
-
-
-def _expand_flat(t, shape):
-    """Broadcast *t* to *shape*, then flatten it.
-
-    Materializes a copy when *t* is genuinely broadcast: the kernels below index
-    every operand at the output element, so they need one element per output.
-    """
-    if tuple(t.shape) != tuple(shape):
-        t = t.expand(shape)
-    return t.contiguous().reshape(-1)
-
-
-def coalesce_broadcast_dims(a_shape, b_shape):
-    """Coalesce N-dim broadcast into minimal effective dimensions.
-
-    Merges adjacent dimensions that have the same broadcast behaviour
-    (both real or both broadcast) to minimise the number of divmod
-    operations inside the kernel loop.
-
-    This is a lowering decision, not part of the op's contract: the Op layer
-    hands ``BinaryKernel`` the two operand shapes and the kernel picks its own
-    index representation.
-
-    Args:
-        a_shape: Shape tuple of input a.
-        b_shape: Shape tuple of input b.
-
-    Returns:
-        Tuple of (out_shape, coalesced_shape, a_strides, b_strides) where
-        strides use 0 for broadcast dimensions.
-    """
-    # Normalise scalar (0-dim) inputs to 1-dim with size 1
-    if len(a_shape) == 0:
-        a_shape = (1,)
-    if len(b_shape) == 0:
-        b_shape = (1,)
-
-    out_shape = torch.broadcast_shapes(a_shape, b_shape)
-    ndim = len(out_shape)
-    a_pad = (1,) * (ndim - len(a_shape)) + tuple(a_shape)
-    b_pad = (1,) * (ndim - len(b_shape)) + tuple(b_shape)
-
-    def _make_strides(padded_shape):
-        strides = [1] * ndim
-        for i in range(ndim - 2, -1, -1):
-            strides[i] = strides[i + 1] * padded_shape[i + 1]
-        # Only zero strides for genuinely broadcast dims (size-1 expanded to >1)
-        return [0 if padded_shape[i] == 1 and out_shape[i] > 1 else strides[i] for i in range(ndim)]
-
-    a_raw = _make_strides(a_pad)
-    b_raw = _make_strides(b_pad)
-
-    # Coalesce adjacent dims with compatible broadcast patterns
-    groups = [(out_shape[0], a_raw[0], b_raw[0])]
-    for i in range(1, ndim):
-        prev_out, prev_as, prev_bs = groups[-1]
-        a_can = (a_raw[i] == 0 and prev_as == 0) or (
-            a_raw[i] != 0 and prev_as == a_raw[i] * out_shape[i]
-        )
-        b_can = (b_raw[i] == 0 and prev_bs == 0) or (
-            b_raw[i] != 0 and prev_bs == b_raw[i] * out_shape[i]
-        )
-        if a_can and b_can:
-            groups[-1] = (prev_out * out_shape[i], a_raw[i], b_raw[i])
-        else:
-            groups.append((out_shape[i], a_raw[i], b_raw[i]))
-
-    # Remove trivial size-1 groups (unless all trivial)
-    groups = [g for g in groups if g[0] > 1] or [(1, 0, 0)]
-    coalesced_shape = tuple(g[0] for g in groups)
-    a_strides = tuple(g[1] for g in groups)
-    b_strides = tuple(g[2] for g in groups)
-    return out_shape, coalesced_shape, a_strides, b_strides
-
-
-def _compute_broadcast_offsets(flat_idx, ndim, divisors, a_strides, b_strides):
-    """Compute a_off and b_off from flat_idx using compile-time unrolled divmod chain.
-
-    All arguments except flat_idx are Python-level constants, so the loop
-    unrolls at kernel build time.
-    """
-    a_off = 0
-    b_off = 0
-    remaining = flat_idx
-    for d in range(ndim - 1):
-        coord = remaining // divisors[d]
-        remaining = remaining % divisors[d]
-        a_off = a_off + coord * a_strides[d]
-        b_off = b_off + coord * b_strides[d]
-    a_off = a_off + remaining * a_strides[ndim - 1]
-    b_off = b_off + remaining * b_strides[ndim - 1]
-    return a_off, b_off
-
-
-def _is_contiguous_same_shape(coalesced_shape, a_strides, b_strides):
-    """Return True when both inputs are contiguous with the same shape (no broadcast)."""
-    return (
-        len(coalesced_shape) == 1
-        and all(s == 1 for s in a_strides)
-        and all(s == 1 for s in b_strides)
-    )
-
-
-@functools.lru_cache(maxsize=32)
-def _make_binary_register_copy(
-    N_total,
-    dtype,
-    op_func,
-    output_dtype=None,
-    threads=256,
-    num_per_thread=8,
-):
-    """Binary register_copy: fragment load -> compute -> fragment store.
-
-    Only available for same-shape contiguous inputs (no broadcast).
-    Uses T.alloc_fragment + T.copy for vectorized 128-bit memory access,
-    giving ~2-3x bandwidth vs scalar access for complex op_funcs that
-    prevent TVM's auto-vectorizer from kicking in.
-    """
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[2])
-    def kernel(threads, num_per_thread):
-        block_size = threads * num_per_thread
-
-        @T.prim_func
-        def main(
-            a: T.Tensor((N_total,), dtype),
-            b: T.Tensor((N_total,), dtype),
-            y: T.Tensor((N_total,), out_dtype),
-        ):
-            with T.Kernel(T.ceildiv(N_total, block_size), threads=threads) as bx:
-                a_reg = T.alloc_fragment((block_size,), dtype)
-                b_reg = T.alloc_fragment((block_size,), dtype)
-                y_reg = T.alloc_fragment((block_size,), out_dtype)
-                T.copy(a[bx * block_size : (bx + 1) * block_size], a_reg)
-                T.copy(b[bx * block_size : (bx + 1) * block_size], b_reg)
-                for i, j in T.Parallel(threads, num_per_thread):
-                    idx = i * num_per_thread + j
-                    y_reg[idx] = op_func(a_reg[idx], b_reg[idx])
-                T.copy(y_reg, y[bx * block_size : (bx + 1) * block_size])
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_binary_direct(
-    N_total,
-    dtype,
-    op_func,
-    coalesced_shape,
-    a_strides,
-    b_strides,
-    a_numel,
-    b_numel,
-    output_dtype=None,
-    threads=256,
-):
-    """Binary direct: 1 element per thread with stride-based broadcast."""
-    out_dtype = output_dtype or dtype
-
-    # Fast path: same-shape contiguous inputs -- skip broadcast machinery
-    if _is_contiguous_same_shape(coalesced_shape, a_strides, b_strides):
-
-        @tilelang.jit(out_idx=[2])
-        def kernel(threads):
-            @T.prim_func
-            def main(
-                a: T.Tensor((N_total,), dtype),
-                b: T.Tensor((N_total,), dtype),
-                y: T.Tensor((N_total,), out_dtype),
-            ):
-                with T.Kernel(T.ceildiv(N_total, threads), threads=threads) as bx:
-                    for i in T.Parallel(threads):
-                        idx = bx * threads + i
-                        y[idx] = op_func(a[idx], b[idx])
-
-            return main
-
-        return kernel
-
-    ndim = len(coalesced_shape)
-    divisors = [1] * ndim
-    for i in range(ndim - 2, -1, -1):
-        divisors[i] = divisors[i + 1] * coalesced_shape[i + 1]
-
-    @tilelang.jit(out_idx=[2])
-    def kernel(threads):
-        @T.prim_func
-        def main(
-            a: T.Tensor((a_numel,), dtype),
-            b: T.Tensor((b_numel,), dtype),
-            y: T.Tensor((N_total,), out_dtype),
-        ):
-            with T.Kernel(T.ceildiv(N_total, threads), threads=threads) as bx:
-                for i in T.Parallel(threads):
-                    flat_idx = bx * threads + i
-                    a_off, b_off = _compute_broadcast_offsets(
-                        flat_idx,
-                        ndim,
-                        divisors,
-                        a_strides,
-                        b_strides,
-                    )
-                    y[flat_idx] = op_func(a[a_off], b[b_off])
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_binary_explicit(
-    N_total,
-    dtype,
-    op_func,
-    coalesced_shape,
-    a_strides,
-    b_strides,
-    a_numel,
-    b_numel,
-    output_dtype=None,
-    threads=256,
-    num_per_thread=8,
-):
-    """Binary explicit_parallel: N elements per thread with stride-based broadcast."""
-    out_dtype = output_dtype or dtype
-
-    # Fast path: same-shape contiguous inputs -- skip broadcast machinery
-    if _is_contiguous_same_shape(coalesced_shape, a_strides, b_strides):
-
-        @tilelang.jit(out_idx=[2])
-        def kernel(threads, num_per_thread):
-            block_size = threads * num_per_thread
-
-            @T.prim_func
-            def main(
-                a: T.Tensor((N_total,), dtype),
-                b: T.Tensor((N_total,), dtype),
-                y: T.Tensor((N_total,), out_dtype),
-            ):
-                with T.Kernel(T.ceildiv(N_total, block_size), threads=threads) as bx:
-                    for i, j in T.Parallel(threads, num_per_thread):
-                        idx = (bx * threads + i) * num_per_thread + j
-                        y[idx] = op_func(a[idx], b[idx])
-
-            return main
-
-        return kernel
-
-    ndim = len(coalesced_shape)
-    divisors = [1] * ndim
-    for i in range(ndim - 2, -1, -1):
-        divisors[i] = divisors[i + 1] * coalesced_shape[i + 1]
-
-    @tilelang.jit(out_idx=[2])
-    def kernel(threads, num_per_thread):
-        block_size = threads * num_per_thread
-
-        @T.prim_func
-        def main(
-            a: T.Tensor((a_numel,), dtype),
-            b: T.Tensor((b_numel,), dtype),
-            y: T.Tensor((N_total,), out_dtype),
-        ):
-            with T.Kernel(T.ceildiv(N_total, block_size), threads=threads) as bx:
-                for i, j in T.Parallel(threads, num_per_thread):
-                    flat_idx = (bx * threads + i) * num_per_thread + j
-                    a_off, b_off = _compute_broadcast_offsets(
-                        flat_idx,
-                        ndim,
-                        divisors,
-                        a_strides,
-                        b_strides,
-                    )
-                    y[flat_idx] = op_func(a[a_off], b[b_off])
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_fused_gated_direct(M, N, dtype, op_func, threads=256, output_dtype=None):
-    """FusedGated direct: 1 element per thread. x[:, :N] is gate, x[:, N:] is value.
-
-    ``op_func(gate, value)`` is the compound operation that applies the
-    activation to *gate* and multiplies by *value*.  For fp8 dtypes the
-    caller wraps it via ``_wrap_fp8_accumulation`` so this factory stays
-    fp8-agnostic.
-
-    Args:
-        output_dtype: TileLang dtype string for the output tensor. Defaults to dtype.
-    """
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[1])
-    def kernel(threads_arg):
-        @T.prim_func
-        def main(x: T.Tensor((M, 2 * N), dtype), y: T.Tensor((M, N), out_dtype)):
-            with T.Kernel(T.ceildiv(N, threads_arg), M, threads=threads_arg) as (bx, by):
-                for i in T.Parallel(threads_arg):
-                    col = bx * threads_arg + i
-                    gate = x[by, col]
-                    value = x[by, N + col]
-                    y[by, col] = op_func(gate, value)
-
-        return main
-
-    return kernel
-
-
-@functools.lru_cache(maxsize=32)
-def _make_fused_gated_explicit(
-    M, N, dtype, op_func, threads=256, num_per_thread=8, output_dtype=None
-):
-    """FusedGated explicit_parallel: N elements per thread.
-
-    ``op_func(gate, value)`` is the compound operation (see
-    ``_make_fused_gated_direct``).  fp8 accumulation belongs to the caller, which
-    wraps ``op_func`` via ``_wrap_fp8_accumulation``.
-
-    The grid comes from the specialized arguments, for the reason
-    :func:`_make_unary_explicit` gives.
-
-    Args:
-        output_dtype: TileLang dtype string for the output tensor. Defaults to dtype.
-    """
-    out_dtype = output_dtype or dtype
-
-    @tilelang.jit(out_idx=[1])
-    def kernel(threads_arg, npt_arg):
-        block_N = threads_arg * npt_arg
-
-        @T.prim_func
-        def main(x: T.Tensor((M, 2 * N), dtype), y: T.Tensor((M, N), out_dtype)):
-            with T.Kernel(T.ceildiv(N, block_N), M, threads=threads_arg) as (bx, by):
-                for i, j in T.Parallel(threads_arg, npt_arg):
-                    col = (bx * threads_arg + i) * npt_arg + j
-                    gate = x[by, col]
-                    value = x[by, N + col]
-                    y[by, col] = op_func(gate, value)
-
-        return main
-
-    return kernel
 
 
 class UnaryKernel(Kernel):
